Extract PoseDisplaySelectionHandler
